feat: Move deprecated APIs to seperate files to cleanup
